requirements.txt
django==4.2
djangorestframework
psycopg2-binary
celery
redis
pandas
openpyxl
gunicorn
